refactor(nbd-cow): split cow bitmap internals
